Model migrations (#33)
* changing project structure, temporary commit

* project architecture altered

* project default tables migration

* travis edited to 80

* minor changes for pylint

* minor changes for pylint

* just for rerunning coveralls

* add trivial testcase

* delete unnecessary import

* minor changes for pylint

* minor changes of .travis.yml

* grant all privileges to mysql user

* modify .travis.yml

* modify test comments

* remove trailing whitespace
