+++ .travis.yml
@@ -40,21 +40,13 @@ matrix:
     before_install:
     - mysql -e "CREATE DATABASE PapersFeed_DB;"
     - mysql -e "CREATE USER 'PapersFeed'@'localhost' IDENTIFIED WITH mysql_native_password BY 'swpp2019team3';"
-    - mysql -e "GRANT ALL PRIVILEGES ON PapersFeed_DB.* to 'PapersFeed'@'localhost';"
+    - mysql -e "GRANT ALL PRIVILEGES ON *.* to 'PapersFeed'@'localhost';"
     install:
     - pip install -r requirements.txt
     script:
     - pylint --rcfile=./backend/.pylintrc backend --disable='fixme, duplicate-code'
-    - coverage run --source='.' ./backend/manage.py test
-    - coverage report --fail-under=90
+    - cd backend
+    - coverage run --source='.' ./manage.py test
+    - coverage report --fail-under=85
     after_success:
     - coveralls
-
-  - dist: trusty
-    addons:
-      sonarcloud:
-        organization: "swsnu"
-        token:
-          secure: "noY4L8OXZLjMCVPaeUernI+gxBcC9Nh8fQbW5MN5dw1UtnJXbLnl+Br1SK6Dip8i+Z7ETSm4QTrkuPyvBSSRfUwJH4jo5cZ+ZA8+d/pKab96iQdUr4riTuEXuT4hKSdj2tdBJtVQLt7CTbGVyJP43UDwJHxDbPIh721sbT4e2DcCoA1D2H2S5OGcbjWdD4Z8av6kPpaeEybCbr22x6nPLfxC/veBul+mEbwZKciiV6NDDbYbgfrq398RMe67ZcmaJoLtOf0WvNeppI3y87o/N2JVQkVbPAz4qCQN68m7KDpU3cA28vrKVBa+xr7a7OSsHdcGwVek7H+Hte5JA8K0slfX/4SbmWTH7R8QzZMnN/LRcwKMJyzQV6icfgNGRvQtdg7z192Krb1IpiCrAKXToXTUNom/RVufqbWc5Mf6KU3sHVrbLJJx7sDXYKY9rhROpdZitdN0fhw3lH3IUByGsDHElMTr4z3aqaYv03OOiQG1R9JsURbEYl1qMb5n/JIl9qtAkFQ/c0pFYa0rEppGzsZwX9yLf8e3XnjkuZXQLmmPkYqD2ix4cksFQMJK94ZjR2hkN/K4Vljz0Z1EAOiFhnGWW/H3tt4Q0z50QB2WY3y+HMzEpLQ11qmKHZ9eT8BYmlDXwglejcC9KTTwICO/TvjM5rPLccYa/hzLY5yEQN0="
-    script:
-    - sonar-scanner
