+++ backend/PapersFeed_backend/settings.py
@@ -32,7 +32,6 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     'papersfeed.apps.PapersfeedConfig',
-    'papers.apps.PapersConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
