--- backend/papers/apis.py
@@ -1,6 +1,0 @@
-"""apis.py"""
-# -*- coding: utf-8 -*-
-
-# Internal Modules
-# from papers.models.paper import *
-# from PapersFeed_backend.utils import *
