--- backend/papers/apps.py
@@ -1,8 +1,0 @@
-"""apps.py"""
-from django.apps import AppConfig
-
-
-class PapersConfig(AppConfig):
-    """Papers Config"""
-    name = 'papers'
-
