--- backend/papers/urls.py
@@ -1,9 +1,0 @@
-"""urls.py"""
-# -*- coding: utf-8 -*-
-# pylint: skip-file
-
-from django.conf.urls import url
-from . import views
-
-urlpatterns = [
-]
