--- backend/papers/views.py
@@ -1,8 +1,0 @@
-"""views.py"""
-# -*- coding: utf-8 -*-
-
-# Python Modules
-#import os
-#import json
-#import datetime
-#import traceback
