+++ backend/papersfeed/constants.py
@@ -15,3 +15,15 @@ DATA = 'data'
 DEBUG = 'debug'
 ERROR = 'error'
 DESCRIPTION = 'description'
+ID = 'id'
+COUNT = 'count'
+SESSION = 'session'
+
+# User
+USER = 'user'
+EMAIL = 'email'
+USERNAME = 'username'
+PASSWORD = 'password'
+IS_FOLLOWING = 'is_following'
+IS_FOLLOWED = 'is_followed'
+FOLLOWER = 'follower'
