--- backend/papersfeed/models.py
@@ -1,1 +1,0 @@
-"""models.py"""
