--- backend/papersfeed/utils.py
@@ -1,7 +1,0 @@
-"""utils.py"""
-# -*- coding: utf-8 -*-
-
-
-class ApiError(Exception):
-    """raise ApiError"""
-    print('API ERROR')
