+++ backend/papersfeed/utils/users/utils.py
@@ -1,0 +1,1 @@
+"""utils.py"""
